+++ src/core/shared/enums/system.ts
@@ -1,4 +1,8 @@
 export enum SYSTEM_EVENTS {
+    APPEND_BLIP = 'append:Blip',
+    APPEND_MARKER = 'append:Marker',
+    APPEND_TEXTLABELS = 'append:TextLabel',
+    //
     BOOTUP_ENABLE_ENTRY = 'enable:Entry',
     //
     INTERACTION = 'player:Interact',
@@ -45,6 +49,10 @@ export enum SYSTEM_EVENTS {
     QUICK_TOKEN_FETCH = 'quicktoken:Fetch',
     QUICK_TOKEN_NONE = 'quicktoken:None',
     QUICK_TOKEN_UPDATE = 'quicktoken:Update',
+    //
+    REMOVE_MARKER = 'remove:Marker',
+    REMOVE_BLIP = 'remove:Blip',
+    REMOVE_TEXTLABEL = 'remove:Textlabel',
     //
     PLAY_PARTICLE_EFFECT = 'ptfx:Play',
     //