+++ src/core/shared/interfaces/TextLabel.ts
@@ -4,4 +4,5 @@ export interface TextLabel {
     pos: Vector3;
     data: string;
     maxDistance?: number;
+    uid?: string;
 }