+++ documentation/changelog.md
@@ -38,6 +38,9 @@ description: List of all the changes and current versions.
 + Prevent Opening Inventory while Dead
 + Added Unique Athena Events for Vehicles, Players
 + Fixed Global Text Labels
++ Added Append & Remove to Textlabels
++ Added Append & Remove to Markers
++ Added Append & Remove to Blips
 ```
 
 ## 1.6.0 - February 1-10, 2021